feat: split up main func
